+++ webots_ros2_tiago/CHANGELOG.rst
@@ -2,9 +2,13 @@
 Changelog for package webots_ros2_tiago
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-2023.0.4 (2023-XX-XX)
+2023.1.0 (2023-XX-XX)
 ------------------
 * Added new world, resources and launch file to start the TIAGo with real robot configuration.
+
+2023.0.4 (2023-XX-XX)
+------------------
+* Start ros control and navigation nodes when Webots is ready.
 
 2023.0.3 (2023-04-12)
 ------------------